refactor: remove LLM interaction skill and update Optimus configurations
