+++ .gitignore
@@ -30,7 +30,6 @@ packages/valory/skills/*
 !packages/valory/services/optimus_pearl
 !packages/valory/skills/liquidity_trader_abci
 !packages/valory/skills/optimus_abci
-!packages/valory/skills/llm_interaction
 
 leak_report
 /Pipfile.lock
