+++ packages/packages.json
@@ -14,11 +14,10 @@
         "contract/valory/staking_token/0.1.0": "bafybeiev3fx5wg4rs5rciqdxemdvt73fwu2nylv23i7g665jzorcyyjape",
         "contract/valory/staking_activity_checker/0.1.0": "bafybeigse23vkx6rxws7ydkl44gtqs5z2g7fo4qejb36pjznegkbv5tkeu",
         "contract/valory/sturdy_yearn_v3_vault/0.1.0": "bafybeif7jichdhvvptgp6ezovod6em47qfid7kr3wtdi77l3ihqbzr5twe",
-        "skill/valory/llm_interaction/0.1.0": "bafybeibrrgso3pecrkh5jhj2en6r4vuk7kycqn3sbma4rl32g4h7zfvr4e",
-        "skill/valory/liquidity_trader_abci/0.1.0": "bafybeigvfuq4bvbo7dz44kkngfmgmjd4ezjnieyd23pxha37keup2ffomy",
-        "skill/valory/optimus_abci/0.1.0": "bafybeic3tzsx4nsf4efikf23ihgpdk4fa5osqvvl2uwf4c7tsq4btrqmra",
-        "agent/valory/optimus/0.1.0": "bafybeid6dlsbzr5ked7pmy4lfuntdfkhzscqjc4ucojsxnsds5l2twxpdq",
-        "service/valory/optimus/0.1.0": "bafybeifaeksyzvbnn7zfn3vphfudup2dw4ncg3jhsckm3q5fxzdxzg3igi"
+        "skill/valory/liquidity_trader_abci/0.1.0": "bafybeigq25mrf4lkfgcimgcyrqinfgrwnydbyhlxddemfheqnbwcanehvm",
+        "skill/valory/optimus_abci/0.1.0": "bafybeiellscd6rktkdqxoycic5q7xh2kqafh4jf3rvwkq32fbmbqgrj2ca",
+        "agent/valory/optimus/0.1.0": "bafybeiciapopdevhjpyp7yvqw7e6de5xyfojobxdqyfj7upxjyr4wqgtku",
+        "service/valory/optimus/0.1.0": "bafybeif6nbkd3ae6tlakwmb4m4o6pfewpqkxbf4j6tfl6k5rarbx7d7zhu"
     },
     "third_party": {
         "protocol/dvilela/kv_store/0.1.0": "bafybeihimf5f37uupxmugvagmaxworgmz7cxuqpikkyzlgldtbq46jbvci",
