+++ packages/valory/agents/optimus/aea-config.yaml
@@ -39,13 +39,12 @@ protocols:
 skills:
 - valory/abstract_abci:0.1.0:bafybeia2egqettv7fd4yryya5ocrxdmpxibvt5ayjg2npvblz73rjpoxla
 - valory/abstract_round_abci:0.1.0:bafybeia27qmw6w5ds5fcrpj2475brnz742aampe3sgochloijs2l7jovai
-- valory/liquidity_trader_abci:0.1.0:bafybeigvfuq4bvbo7dz44kkngfmgmjd4ezjnieyd23pxha37keup2ffomy
-- valory/optimus_abci:0.1.0:bafybeic3tzsx4nsf4efikf23ihgpdk4fa5osqvvl2uwf4c7tsq4btrqmra
+- valory/liquidity_trader_abci:0.1.0:bafybeigq25mrf4lkfgcimgcyrqinfgrwnydbyhlxddemfheqnbwcanehvm
+- valory/optimus_abci:0.1.0:bafybeiellscd6rktkdqxoycic5q7xh2kqafh4jf3rvwkq32fbmbqgrj2ca
 - valory/registration_abci:0.1.0:bafybeifaa6ejpihwxxxswqd5qfy63rixml43d3ljnezilflvdtclspjj6y
 - valory/reset_pause_abci:0.1.0:bafybeigebq46oqz2mx2iajupr6p5pgm6z5pvfye5w6zypsseuqtvta7b4a
 - valory/termination_abci:0.1.0:bafybeieurwmfernodqyczj5ertsgfbjtjnrlgvte7sli4sajnbopty7inu
 - valory/transaction_settlement_abci:0.1.0:bafybeifkftgkyzrxwxjdyqixpp7vk6aqmufikalmwx3kydtlg74tonu47u
-- valory/llm_interaction:0.1.0:bafybeibrrgso3pecrkh5jhj2en6r4vuk7kycqn3sbma4rl32g4h7zfvr4e
 customs:
 - valory/merkl_pools_search:0.1.0:bafybeihg3shkrvx7yzrqenjow23ebcl5xskrzrqkj47er2z47xhhs2z6ma
 - valory/max_apr_selection:0.1.0:bafybeidj3qjcq5zrtysjhgw4f2odxnr66ij4od6oluh655v3ln5uxcepci
