+++ packages/valory/services/optimus/service.yaml
@@ -6,7 +6,7 @@ aea_version: '>=1.0.0, <2.0.0'
 license: Apache-2.0
 fingerprint: {}
 fingerprint_ignore_patterns: []
-agent: valory/optimus:0.1.0:bafybeid6dlsbzr5ked7pmy4lfuntdfkhzscqjc4ucojsxnsds5l2twxpdq
+agent: valory/optimus:0.1.0:bafybeiciapopdevhjpyp7yvqw7e6de5xyfojobxdqyfj7upxjyr4wqgtku
 number_of_agents: 1
 deployment:
   agent:
