--- packages/valory/skills/llm_interaction/skill.yaml
@@ -1,90 +1,0 @@
-name: llm_interaction
-author: valory
-version: 0.1.0
-type: skill
-description: The scaffold skill is a scaffold for your own skill implementation.
-license: Apache-2.0
-aea_version: '>=1.0.0, <2.0.0'
-fingerprint:
-  __init__.py: bafybeiaokwzvrvgjvzf53ovuhew46w4soo6ydfqrmrnwdud4wr36zla4zi
-  behaviours.py: bafybeiglozbntfit6rrxda5dhfpm5itklpy7vv4shui2uw33zqiv6tcf3a
-  dialogues.py: bafybeiduwarr6zlueylecydixshj2uqdojjbke3oqlocmyoajn3qurtyyq
-  handlers.py: bafybeiefqxudty4zlynh44uagqokyznst6hyq3jy7bm7ykbztbzy347cwe
-  models.py: bafybeiawpn7ptjwnqiacdccfsjdl62s354zetmhsyxodbco2464wxknzxa
-  modius-ui-build/antd-reset.css: bafybeifyeozf54phizmvhn5cfvvsn6vsnceqje5vlm2hwaiw5rtnmsuahe
-  modius-ui-build/asset-manifest.json: bafybeidwmfoahxhjl3sayuc7jvoe22lu2qculqbsannlvl74lcbfprku2a
-  modius-ui-build/favicon.ico: bafybeigoxdaiyjbhlh3hq7b5myaowqlc22poxirhfjqzesxk4erea2e3pu
-  modius-ui-build/globals.css: bafybeidc6ngo76cbhkhhhcddqw4g5id6rfabp723qwqq6y362d6xc3qahy
-  modius-ui-build/index.html: bafybeidbw7bisw5wlsucra4gmnuasftams7tim3m3ke6iwq467ipbpcyjq
-  modius-ui-build/logos/modius.png: bafybeih4k42vb6nhnbrez2pc6gfllraupihnwx5y6w4iuupw2rwm6hehty
-  modius-ui-build/logos/protocols/balancer.png: bafybeicqmyeayxigss5t2tpbfad5dl4yfrd6d6sqndrgdnt33jwnskicnm
-  modius-ui-build/logos/protocols/mode-network.png: bafybeieacbmo2f6r7uc6i7ftvw6vl2epx4k3gc5rubz7z46ug6q6xc244a
-  modius-ui-build/logos/protocols/sturdy.png: bafybeihymprfge4jyppwzsykd7btospy7q4ofswrd3nxx6qrpjxfbqvrse
-  modius-ui-build/logos/tokens/bal.png: bafybeidnd5afsjcevrmyd4ucf74auemv5c7lzyurieemmdv4ht4idzp6qi
-  modius-ui-build/logos/tokens/dai.png: bafybeifysdjm4t6mnzqvwrh5snqxbsiubvoa2akizyb5oiqaluxpwpqu7i
-  modius-ui-build/logos/tokens/eth.png: bafybeiejj4gaj2ahgr6kavbg5rd5cnrpuxaicobfcd4ffxmlob664ouyrq
-  modius-ui-build/logos/tokens/ezeth.png: bafybeigknvkgjfidvpr2mkj7pfpt5dy2ecbn7z2kosdo2xc53crqh2bqe4
-  modius-ui-build/logos/tokens/mode.png: bafybeiakyu4lwlzqgjluol6gpfbydezwh7cosj5mncpmorbwbe46c6ck3a
-  modius-ui-build/logos/tokens/olas.png: bafybeiguddvyhbusz6hya4rysexr45sn2nk5yn7mlwf34qdfrknxzxgfmi
-  modius-ui-build/logos/tokens/op.png: bafybeiejhf2onv6i3gjj4busavchwr2246br277bi7t5r2yugsfgkiou3e
-  modius-ui-build/logos/tokens/pxeth.png: bafybeidzbxp6eu3577tdhyjtxpezwegdi2l6norylg3lp2lzkkpd5rzoiq
-  modius-ui-build/logos/tokens/reth.png: bafybeihgvupiwdhxy75id46lgd6o4kccppgsixveglymjtiefditsa2b2e
-  modius-ui-build/logos/tokens/steth.png: bafybeifdxeuui5pjngbeku6rjgxewxieqcxtr2ggwmbmrn7pxhu3gt6apu
-  modius-ui-build/logos/tokens/stone.png: bafybeia255ikjhe4azvdc6d6bnndfdp3jtduzr7jwibacvvpp6o5qxx5jq
-  modius-ui-build/logos/tokens/tbtc.png: bafybeigmbnetcoo7yfacmxmzfej4aerqg4clpvabekfjdl7doaorrygley
-  modius-ui-build/logos/tokens/usdc.png: bafybeiddabigc6c5ga27grotfy2bfr247eah2qs3nsjzup7z3l4r2kvggm
-  modius-ui-build/logos/tokens/usdt.png: bafybeibx7p2cckrjejds6cwvu23wco5uqgnlmqmdumxytlyqarx5v4j2zy
-  modius-ui-build/logos/tokens/weth.png: bafybeie6swfujvdlkgjtg3vbjejhkiv6brb3zpjxqrvquzdkci375fpjde
-  modius-ui-build/manifest.json: bafybeial7amzp5ze4y773v7i67igygm3evylffeqxwr56aw3332ioavks4
-  modius-ui-build/robots.txt: bafybeifmudop7z45oyjc7jimzav6aqogvpwkobbtscdwidtpuovh6yp5wu
-  modius-ui-build/static/js/main.d1640d96.js: bafybeihfcayc4q7n6h7dvletzukoudzhfmwqtinsx4vzgnnwufr7jktvey
-  modius-ui-build/static/js/main.d1640d96.js.LICENSE.txt: bafybeig2lpx45vgz5yzhqrphmzyy3yhja6tfq76mvhkssx7tpeiwgd6kfy
-  modius-ui-build/static/js/main.d1640d96.js.map: bafybeifmueqrp5566xpvehfma54wk7dbx53zf5icwwakjflxkc3y6wtn2i
-  prompts.py: bafybeif22rfwhd45kipduh6eapldfzcfxcm2ux3ezqudejb3pbykjhd6eq
-fingerprint_ignore_patterns: []
-connections:
-- valory/http_server:0.22.0:bafybeic3jpkum7g6qo6x6vdrmvvhj7vqw7ec2op72uc3yfhmnlp5hn3joy
-- valory/openai:0.1.0:bafybeiaryip2lbiluf7vjsdq6r5qiqdesmhqevrzekfrxpugfypqja6gae
-- dvilela/kv_store:0.1.0:bafybeiekugvb2kan4342hliluxl3h5och3fjwqoafdyttpcn57evvyztq4
-contracts: []
-protocols:
-- valory/http:1.0.0:bafybeih4azmfwtamdbkhztkm4xitep3gx6tfdnoz6tvllmaqnhu3klejfa
-- valory/llm:1.0.0:bafybeihj2orqapokafg7co3wlogwgf6cckaaj2bz3nuv4rm3opkoyhrlp4
-- dvilela/kv_store:0.1.0:bafybeihimf5f37uupxmugvagmaxworgmz7cxuqpikkyzlgldtbq46jbvci
-skills:
-- valory/abstract_round_abci:0.1.0:bafybeia27qmw6w5ds5fcrpj2475brnz742aampe3sgochloijs2l7jovai
-- valory/liquidity_trader_abci:0.1.0:bafybeigvfuq4bvbo7dz44kkngfmgmjd4ezjnieyd23pxha37keup2ffomy
-behaviours:
-  llm_interaction:
-    args: {}
-    class_name: LlmInteractionBehaviour
-handlers:
-  http_handler:
-    args: {}
-    class_name: HttpHandler
-  kv_store:
-    args: {}
-    class_name: KvStoreHandler
-  llm_handler:
-    args: {}
-    class_name: LlmHandler
-models:
-  kv_store_dialogues:
-    args: {}
-    class_name: KvStoreDialogues
-  llm_dialogues:
-    args: {}
-    class_name: LlmDialogues
-  http_dialogues:
-    args: {}
-    class_name: HttpDialogues
-  params:
-    args: {}
-    class_name: LlmInteractionParams
-  requests:
-    args: {}
-    class_name: Requests
-dependencies:
-  pyyaml:
-    version: ==6.0.1
-is_abstract: false
