+++ packages/valory/skills/optimus_abci/dialogues.py
@@ -73,10 +73,6 @@ AbciDialogue = BaseAbciDialogue
 AbciDialogues = BaseAbciDialogues
 
 
-HttpDialogue = BaseHttpDialogue
-HttpDialogues = BaseHttpDialogues
-
-
 SigningDialogue = BaseSigningDialogue
 SigningDialogues = BaseSigningDialogues
 
@@ -96,5 +92,83 @@ TendermintDialogues = BaseTendermintDialogues
 IpfsDialogue = BaseIpfsDialogue
 IpfsDialogues = BaseIpfsDialogues
 
+
+from typing import Any
+from aea.protocols.base import Address, Message
+from aea.protocols.dialogue.base import Dialogue as BaseDialogue
+from aea.skills.base import Model
+from packages.dvilela.protocols.kv_store.dialogues import (
+    KvStoreDialogue as BaseKvStoreDialogue,
+)
+from packages.dvilela.protocols.kv_store.dialogues import (
+    KvStoreDialogues as BaseKvStoreDialogues,
+)
+from packages.valory.protocols.llm.dialogues import LlmDialogue as BaseLlmDialogue
+from packages.valory.protocols.llm.dialogues import LlmDialogues as BaseLlmDialogues
+from packages.valory.skills.abstract_round_abci.dialogues import (
+    HttpDialogue as BaseHttpDialogue,
+)
+from packages.valory.skills.abstract_round_abci.dialogues import (
+    HttpDialogues as BaseHttpDialogues,
+)
+
+LlmDialogue = BaseLlmDialogue
+HttpDialogue = BaseHttpDialogue
+HttpDialogues = BaseHttpDialogues
 KvStoreDialogue = BaseKvStoreDialogue
-KvStoreDialogues = BaseKvStoreDialogues
+
+class LlmDialogues(Model, BaseLlmDialogues):
+    """A class to keep track of LLM dialogues."""
+
+    def __init__(self, **kwargs: Any) -> None:
+        """
+        Initialize dialogues.
+
+        :param kwargs: keyword arguments
+        """
+        Model.__init__(self, **kwargs)
+
+        def role_from_first_message(  # pylint: disable=unused-argument
+            message: Message, receiver_address: Address
+        ) -> BaseDialogue.Role:
+            """Infer the role of the agent from an incoming/outgoing first message
+
+            :param message: an incoming/outgoing first message
+            :param receiver_address: the address of the receiving agent
+            :return: The role of the agent
+            """
+            return LlmDialogue.Role.SKILL
+
+        BaseLlmDialogues.__init__(
+            self,
+            self_address=str(self.skill_id),
+            role_from_first_message=role_from_first_message,
+        )
+
+class KvStoreDialogues(Model, BaseKvStoreDialogues):
+    """The dialogues class keeps track of all dialogues."""
+
+    def __init__(self, **kwargs: Any) -> None:
+        """
+        Initialize dialogues.
+
+        :param kwargs: keyword arguments
+        """
+        Model.__init__(self, **kwargs)
+
+        def role_from_first_message(  # pylint: disable=unused-argument
+            message: Message, receiver_address: Address
+        ) -> BaseDialogue.Role:
+            """Infer the role of the agent from an incoming/outgoing first message
+
+            :param message: an incoming/outgoing first message
+            :param receiver_address: the address of the receiving agent
+            :return: The role of the agent
+            """
+            return KvStoreDialogue.Role.SKILL
+
+        BaseKvStoreDialogues.__init__(
+            self,
+            self_address=str(self.skill_id),
+            role_from_first_message=role_from_first_message,
+        )
