+++ packages/valory/skills/optimus_abci/modius-ui-build/asset-manifest.json
