+++ packages/valory/skills/optimus_abci/modius-ui-build/favicon.ico
