+++ packages/valory/skills/optimus_abci/modius-ui-build/logos/tokens/mode.png
