+++ packages/valory/skills/optimus_abci/modius-ui-build/manifest.json
