+++ packages/valory/skills/optimus_abci/skill.yaml
@@ -9,21 +9,56 @@ fingerprint:
   __init__.py: bafybeiechr3zr5bc4xl3vvs2p2gti54ily7ao2gu4ff5lys6cixehzkdea
   behaviours.py: bafybeig3agtm256pl7rekx3i3kbz6h5hrnneu44f4pzkqp4mwtdxuob54e
   composition.py: bafybeih3laiba2txmxtw43zhtbuppc3gzthbb7osejvf5ngf5ioyn22akm
-  dialogues.py: bafybeihuayabnubriylso5vlvj2qgncfb66ecaizzoxiob3r5airi7mexa
+  dialogues.py: bafybeidyrestjlqswuwasecfwmgle5qltamlg5aiinyujisrgfp2j5qzbe
   fsm_specification.yaml: bafybeicbqp4idp4ndsh2ffmmfjbn7e3b7dizt2dyrwiupzosg2d4drsupm
-  handlers.py: bafybeihwk4gfdhzpsexwllt6ujzf67t4pmqzl353piiuepqjgakeij3ob4
-  models.py: bafybeighgbho762qt4czzvktvkucu4mv6hzungwv6tw3tvujsshi2q2zhm
+  handlers.py: bafybeid2xit5ueixrrk7ecpnk3bz4ug4kisp5x35hqvuuvtklfdwpm2rjy
+  models.py: bafybeicpylxjha54nt36nvohc3hfx5esxi25wsupsqmma5lu7wr2wt2a4a
+  modius-ui-build/antd-reset.css: bafybeifyeozf54phizmvhn5cfvvsn6vsnceqje5vlm2hwaiw5rtnmsuahe
+  modius-ui-build/asset-manifest.json: bafybeidwmfoahxhjl3sayuc7jvoe22lu2qculqbsannlvl74lcbfprku2a
+  modius-ui-build/favicon.ico: bafybeigoxdaiyjbhlh3hq7b5myaowqlc22poxirhfjqzesxk4erea2e3pu
+  modius-ui-build/globals.css: bafybeidc6ngo76cbhkhhhcddqw4g5id6rfabp723qwqq6y362d6xc3qahy
+  modius-ui-build/index.html: bafybeidbw7bisw5wlsucra4gmnuasftams7tim3m3ke6iwq467ipbpcyjq
+  modius-ui-build/logos/modius.png: bafybeih4k42vb6nhnbrez2pc6gfllraupihnwx5y6w4iuupw2rwm6hehty
+  modius-ui-build/logos/protocols/balancer.png: bafybeicqmyeayxigss5t2tpbfad5dl4yfrd6d6sqndrgdnt33jwnskicnm
+  modius-ui-build/logos/protocols/mode-network.png: bafybeieacbmo2f6r7uc6i7ftvw6vl2epx4k3gc5rubz7z46ug6q6xc244a
+  modius-ui-build/logos/protocols/sturdy.png: bafybeihymprfge4jyppwzsykd7btospy7q4ofswrd3nxx6qrpjxfbqvrse
+  modius-ui-build/logos/tokens/bal.png: bafybeidnd5afsjcevrmyd4ucf74auemv5c7lzyurieemmdv4ht4idzp6qi
+  modius-ui-build/logos/tokens/dai.png: bafybeifysdjm4t6mnzqvwrh5snqxbsiubvoa2akizyb5oiqaluxpwpqu7i
+  modius-ui-build/logos/tokens/eth.png: bafybeiejj4gaj2ahgr6kavbg5rd5cnrpuxaicobfcd4ffxmlob664ouyrq
+  modius-ui-build/logos/tokens/ezeth.png: bafybeigknvkgjfidvpr2mkj7pfpt5dy2ecbn7z2kosdo2xc53crqh2bqe4
+  modius-ui-build/logos/tokens/mode.png: bafybeiakyu4lwlzqgjluol6gpfbydezwh7cosj5mncpmorbwbe46c6ck3a
+  modius-ui-build/logos/tokens/olas.png: bafybeiguddvyhbusz6hya4rysexr45sn2nk5yn7mlwf34qdfrknxzxgfmi
+  modius-ui-build/logos/tokens/op.png: bafybeiejhf2onv6i3gjj4busavchwr2246br277bi7t5r2yugsfgkiou3e
+  modius-ui-build/logos/tokens/pxeth.png: bafybeidzbxp6eu3577tdhyjtxpezwegdi2l6norylg3lp2lzkkpd5rzoiq
+  modius-ui-build/logos/tokens/reth.png: bafybeihgvupiwdhxy75id46lgd6o4kccppgsixveglymjtiefditsa2b2e
+  modius-ui-build/logos/tokens/steth.png: bafybeifdxeuui5pjngbeku6rjgxewxieqcxtr2ggwmbmrn7pxhu3gt6apu
+  modius-ui-build/logos/tokens/stone.png: bafybeia255ikjhe4azvdc6d6bnndfdp3jtduzr7jwibacvvpp6o5qxx5jq
+  modius-ui-build/logos/tokens/tbtc.png: bafybeigmbnetcoo7yfacmxmzfej4aerqg4clpvabekfjdl7doaorrygley
+  modius-ui-build/logos/tokens/usdc.png: bafybeiddabigc6c5ga27grotfy2bfr247eah2qs3nsjzup7z3l4r2kvggm
+  modius-ui-build/logos/tokens/usdt.png: bafybeibx7p2cckrjejds6cwvu23wco5uqgnlmqmdumxytlyqarx5v4j2zy
+  modius-ui-build/logos/tokens/weth.png: bafybeie6swfujvdlkgjtg3vbjejhkiv6brb3zpjxqrvquzdkci375fpjde
+  modius-ui-build/manifest.json: bafybeial7amzp5ze4y773v7i67igygm3evylffeqxwr56aw3332ioavks4
+  modius-ui-build/robots.txt: bafybeifmudop7z45oyjc7jimzav6aqogvpwkobbtscdwidtpuovh6yp5wu
+  modius-ui-build/static/js/main.d1640d96.js: bafybeihfcayc4q7n6h7dvletzukoudzhfmwqtinsx4vzgnnwufr7jktvey
+  modius-ui-build/static/js/main.d1640d96.js.LICENSE.txt: bafybeig2lpx45vgz5yzhqrphmzyy3yhja6tfq76mvhkssx7tpeiwgd6kfy
+  modius-ui-build/static/js/main.d1640d96.js.map: bafybeifmueqrp5566xpvehfma54wk7dbx53zf5icwwakjflxkc3y6wtn2i
+  prompts.py: bafybeicf422nzsqevscledpwjktllleglahmka3m2gcf6l4iuuwjubleq4
 fingerprint_ignore_patterns: []
-connections: []
+connections:
+- valory/http_server:0.22.0:bafybeic3jpkum7g6qo6x6vdrmvvhj7vqw7ec2op72uc3yfhmnlp5hn3joy
+- valory/openai:0.1.0:bafybeiaryip2lbiluf7vjsdq6r5qiqdesmhqevrzekfrxpugfypqja6gae
+- dvilela/kv_store:0.1.0:bafybeiekugvb2kan4342hliluxl3h5och3fjwqoafdyttpcn57evvyztq4
 contracts: []
-protocols: []
+protocols:
+- valory/http:1.0.0:bafybeih4azmfwtamdbkhztkm4xitep3gx6tfdnoz6tvllmaqnhu3klejfa
+- valory/llm:1.0.0:bafybeihj2orqapokafg7co3wlogwgf6cckaaj2bz3nuv4rm3opkoyhrlp4
+- dvilela/kv_store:0.1.0:bafybeihimf5f37uupxmugvagmaxworgmz7cxuqpikkyzlgldtbq46jbvci
 skills:
 - valory/abstract_round_abci:0.1.0:bafybeia27qmw6w5ds5fcrpj2475brnz742aampe3sgochloijs2l7jovai
 - valory/registration_abci:0.1.0:bafybeifaa6ejpihwxxxswqd5qfy63rixml43d3ljnezilflvdtclspjj6y
 - valory/reset_pause_abci:0.1.0:bafybeigebq46oqz2mx2iajupr6p5pgm6z5pvfye5w6zypsseuqtvta7b4a
 - valory/termination_abci:0.1.0:bafybeieurwmfernodqyczj5ertsgfbjtjnrlgvte7sli4sajnbopty7inu
-- valory/liquidity_trader_abci:0.1.0:bafybeigvfuq4bvbo7dz44kkngfmgmjd4ezjnieyd23pxha37keup2ffomy
-- valory/llm_interaction:0.1.0:bafybeibrrgso3pecrkh5jhj2en6r4vuk7kycqn3sbma4rl32g4h7zfvr4e
+- valory/liquidity_trader_abci:0.1.0:bafybeigq25mrf4lkfgcimgcyrqinfgrwnydbyhlxddemfheqnbwcanehvm
 - valory/transaction_settlement_abci:0.1.0:bafybeifkftgkyzrxwxjdyqixpp7vk6aqmufikalmwx3kydtlg74tonu47u
 behaviours:
   main:
@@ -54,6 +89,9 @@ handlers:
   tendermint:
     args: {}
     class_name: TendermintHandler
+  llm_handler:
+    args: {}
+    class_name: LlmHandler
 models:
   abci_dialogues:
     args: {}
@@ -199,6 +237,7 @@ models:
       pnl_check_interval: 10
       available_strategies: null
       default_acceptance_time: null
+      cleanup_freq: 50
     class_name: Params
   coingecko:
     args:
@@ -237,6 +276,9 @@ models:
   kv_store_dialogues:
     args: {}
     class_name: KvStoreDialogues
+  llm_dialogues:
+    args: {}
+    class_name: LlmDialogues
 dependencies:
   PyYAML:
     version: '>=3.10,<=6.0.1'
